commit 20131104a

+++ Record.java
@@ -8,11 +8,11 @@ public class Record {
     public float relativePosition; // relative position in document (0 to 1)
     public int numChars; // number of chars in phrase (including whitespaces)
     public int numWords; // number of words in phrase
-    public int TF; // term-frequency
-    public int TFIDF; // term frequency-inverse document frequency
+    public double TF; // term-frequency
+    public double TFIDF; // term frequency-inverse document frequency
     public int label; // class label {-1,0,1} for unknown, no, yes, respectively 
 	
-    public Record(String i_phrase, int i_keyphraseness, int i_absPosition, float i_relativePosition, int i_numChars, int i_numWords, int i_TF, int i_TFIDF, int i_label) {
+    public Record(String i_phrase, int i_keyphraseness, int i_absPosition, float i_relativePosition, int i_numChars, int i_numWords, double i_TF, double i_TFIDF, int i_label) {
     	phrase = i_phrase;
         keyphraseness = i_keyphraseness;
     	absPosition = i_absPosition;
